Former-commit-id: 36a8134ffc75d61921a6a64db0c2fa455044539f
Former-commit-id: ae1a906aba95e483e466f2c84fafeb0f2ab17571

+++ backend/src/main/java/com/ssafy/i5i/hotelAPI/domain/hotel/controller/AccommodationController.java
@@ -20,14 +20,14 @@ public class AccommodationController {
     private final AccommodationService accommodationService;
 
     @GetMapping("/{name}/{distance}/{sorted}")
-    public List<AccommodationResponseDto> getAccommodationByName(@PathVariable("name") String name, @PathVariable("distance") double distance, @PathVariable("sorted") String sorted){
+    public List<AccommodationResponseDto> getAccommodationByName(@PathVariable("name") String name, @PathVariable("distance") int distance, @PathVariable("sorted") String sorted){
         AttractionNameRequestDto attractionNameRequestDto = new AttractionNameRequestDto(name, distance, sorted);
         System.out.println("!!!!");
         return accommodationService.getAccommodationByName(attractionNameRequestDto);
     }
 
     @GetMapping("/{latitude}/{longitude}/{distance}/{sorted}")
-    public List<AccommodationResponseDto> getAccomodationByCoordinate(@PathVariable("latitude") double latitude, @PathVariable("longitude") double longitude, @PathVariable("distance") long distance, @PathVariable("sorted") String sorted){
+    public List<AccommodationResponseDto> getAccomodationByCoordinate(@PathVariable("latitude") double latitude, @PathVariable("longitude") double longitude, @PathVariable("distance") int distance, @PathVariable("sorted") String sorted){
         AttractionCoordinateRequestDto attractionCoordinateRequestDto = new AttractionCoordinateRequestDto(latitude, longitude, distance, sorted);
 
         return accommodationService.getAccommodationByCoordinate(attractionCoordinateRequestDto);
